refactor: move change stream tests to integration
